move view components to sub dir
